implement new build system for newer modeling
