+++ bob/core/src/util.cc
@@ -1,4 +1,4 @@
-#include <bob/util.h>
+#include <core/util.h>
 
 #include <sstream>
 #include <fstream>
