+++ bob/neural/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.9 FATAL_ERROR)
-project(neural VERSION 1.0.0 LANGUAGES CXX)
+project(BobNeural VERSION 1.0.0 LANGUAGES CXX)
 
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
@@ -12,9 +12,9 @@ find_package(Eigen3 3.3 REQUIRED NO_MODULE)
 
 include_directories("${PROJECT_SOURCE_DIR}/include")
 
-add_executable("${PROJECT_NAME}"
-`"${PROJECT_SOURCE_DIR}/src/main.cc"
- "${PROJECT_SOURCE_DIR}"
+add_executable(${PROJECT_NAME}
+  "${PROJECT_SOURCE_DIR}/src/main.cc"
+  "${PROJECT_SOURCE_DIR}/src/neural.cc"
 )
 
 target_link_libraries(
