+++ bob/neural/src/main.cc
@@ -1,0 +1,6 @@
+#include <iostream>
+
+int main() {
+  std::cout << "Hi" << std::endl;
+  return EXIT_SUCCESS;
+}
